+++ allure_results/006b22af-1025-4a32-ac46-be46c5e9e7b9-container.json
@@ -1,0 +1,1 @@
+{"uuid": "32d4313b-78b8-4ad6-b1f7-739527459aea", "children": ["ccdf7eff-64db-4065-a0b3-b0ec337df4b5"], "befores": [{"name": "browser", "status": "passed", "start": 1716199859721, "stop": 1716199862990}], "afters": [{"name": "browser::0", "status": "passed", "start": 1716199867009, "stop": 1716199867399}], "start": 1716199859721, "stop": 1716199867399}
