+++ allure_results/07d04845-d59c-4c2e-9059-7b1b04c36148-result.json
@@ -1,0 +1,1 @@
+{"name": "Проверка вопросов и ответов на главной", "status": "passed", "description": "Проверка, что ответы соответствуют заявленным", "steps": [{"name": "Нажать на кнопку «Да все привыкли»", "status": "passed", "steps": [{"name": "Нажать на элемент", "status": "passed", "parameters": [{"name": "locator", "value": "['id', 'rcc-confirm-button']"}], "start": 1716199835196, "stop": 1716199835446}], "start": 1716199835196, "stop": 1716199835447}, {"name": "Проскроллить до блока «Вопросы о важном»", "status": "passed", "steps": [{"name": "Проскроллить до элемента", "status": "passed", "parameters": [{"name": "locator", "value": "['xpath', \".//div[text()='Вопросы о важном']\"]"}], "start": 1716199835447, "stop": 1716199835454}, {"name": "Подождать отображение элемента", "status": "passed", "parameters": [{"name": "locator", "value": "['xpath', \".//div[text()='Вопросы о важном']\"]"}], "start": 1716199835454, "stop": 1716199835466}], "start": 1716199835447, "stop": 1716199835466}, {"name": "Нажать на вопрос", "status": "passed", "steps": [{"name": "Нажать на элемент", "status": "passed", "parameters": [{"name": "locator", "value": "('id', 'accordion__heading-2')"}], "start": 1716199835466, "stop": 1716199835693}, {"name": "Подождать отображение элемента", "status": "passed", "parameters": [{"name": "locator", "value": "('id', 'accordion__heading-2')"}], "start": 1716199835693, "stop": 1716199835707}], "parameters": [{"name": "num", "value": "2"}], "start": 1716199835466, "stop": 1716199835707}, {"name": "Сравнить тексты ответов", "status": "passed", "steps": [{"name": "Подождать отображение элемента", "status": "passed", "parameters": [{"name": "locator", "value": "('id', 'accordion__panel-2')"}], "start": 1716199835707, "stop": 1716199835718}], "parameters": [{"name": "num", "value": "2"}], "start": 1716199835707, "stop": 1716199835726}], "parameters": [{"name": "num", "value": "2"}, {"name": "result", "value": "'Допустим, вы оформляете заказ на 8 мая. Мы привозим самокат 8 мая в течение дня. Отсчёт времени аренды начинается с момента, когда вы оплатите заказ курьеру. Если мы привезли самокат 8 мая в 20:30, суточная аренда закончится 9 мая в 20:30.'"}], "start": 1716199835195, "stop": 1716199835727, "uuid": "f31f5fcc-0680-4912-a38f-ec658a89a725", "historyId": "4b9c62f605e37bbd4806fa6731605bc5", "testCaseId": "288451571851ac6be9a051ff55691f53", "fullName": "tests.test_main_page.TestMainPage#test_check_questions_and_answers", "labels": [{"name": "parentSuite", "value": "tests"}, {"name": "suite", "value": "test_main_page"}, {"name": "subSuite", "value": "TestMainPage"}, {"name": "host", "value": "MacBook-Air-Artem.local"}, {"name": "thread", "value": "13739-MainThread"}, {"name": "framework", "value": "pytest"}, {"name": "language", "value": "cpython3"}, {"name": "package", "value": "tests.test_main_page"}]}
