+++ allure_results/0c09bb96-1bbb-48a9-bc2c-942e33cb2425-container.json
@@ -1,0 +1,1 @@
+{"uuid": "a131cf7b-fd64-4262-97b2-aad5060f1eae", "children": ["66970d53-10a6-4970-9c15-6ed2ca6feb1f"], "befores": [{"name": "browser", "status": "passed", "start": 1716199825952, "stop": 1716199828262}], "afters": [{"name": "browser::0", "status": "passed", "start": 1716199828790, "stop": 1716199829169}], "start": 1716199825952, "stop": 1716199829169}
