+++ allure_results/0c6d7075-6131-4280-8639-4d14971139ad-result.json
@@ -1,0 +1,1 @@
+{"name": "Клик по лого «Яндекс»", "status": "passed", "description": "Проверка, что клик по лого «Яндекс» перенаправляет на новую вкладку «Дзен»", "steps": [{"name": "Нажать на лого «Яндекс»", "status": "passed", "steps": [{"name": "Нажать на элемент", "status": "passed", "parameters": [{"name": "locator", "value": "['xpath', \".//img[@alt='Yandex']\"]"}], "start": 1716199820935, "stop": 1716199821197}], "start": 1716199820935, "stop": 1716199821197}, {"name": "Проверить, что открылась страница «Дзен»", "status": "passed", "steps": [{"name": "Перейти на другую вкладку", "status": "passed", "start": 1716199821197, "stop": 1716199821200}, {"name": "Подождать отображение элемента", "status": "passed", "parameters": [{"name": "locator", "value": "['id', 'dzen-header']"}], "start": 1716199821200, "stop": 1716199822268}], "start": 1716199821197, "stop": 1716199822268}], "start": 1716199820934, "stop": 1716199822268, "uuid": "2750e31b-53b0-4e97-8f94-a848cc52d96d", "historyId": "e2f7a3aaa419ef5d9fd48d008bcae06d", "testCaseId": "e2f7a3aaa419ef5d9fd48d008bcae06d", "fullName": "tests.test_main_page.TestMainPage#test_click_logo_ya_redirect_dzen", "labels": [{"name": "parentSuite", "value": "tests"}, {"name": "suite", "value": "test_main_page"}, {"name": "subSuite", "value": "TestMainPage"}, {"name": "host", "value": "MacBook-Air-Artem.local"}, {"name": "thread", "value": "13739-MainThread"}, {"name": "framework", "value": "pytest"}, {"name": "language", "value": "cpython3"}, {"name": "package", "value": "tests.test_main_page"}]}
