+++ allure_results/0e70d361-f206-4e81-b26f-20bb8503d3d0-container.json
@@ -1,0 +1,1 @@
+{"uuid": "e0ead18a-290e-45a2-973d-8b921b217a99", "befores": [{"name": "result", "status": "passed", "start": 1716199838869, "stop": 1716199838870}], "start": 1716199838869, "stop": 1716199839410}
