+++ allure_results/1abb1ad1-86c1-4a2b-aaed-bd63b438d5ce-container.json
@@ -1,0 +1,1 @@
+{"uuid": "ee458052-00d8-4241-8c18-84124da5a5a3", "befores": [{"name": "result", "status": "passed", "start": 1716199835195, "stop": 1716199835195}], "start": 1716199835195, "stop": 1716199835727}
