+++ allure_results/1f68a72f-a50d-4bb9-be72-a8c62b480df3-container.json
@@ -1,0 +1,1 @@
+{"uuid": "985ed97b-efce-4e8e-a29c-c3541962878d", "children": ["1b07fbd2-d8d0-4816-9a7a-a77ff89255b4"], "befores": [{"name": "browser", "status": "passed", "start": 1716199846225, "stop": 1716199848725}], "afters": [{"name": "browser::0", "status": "passed", "start": 1716199849254, "stop": 1716199849626}], "start": 1716199846225, "stop": 1716199849626}
