+++ allure_results/300411e0-3440-4e22-9f57-838ee7c7a329-result.json
@@ -1,0 +1,1 @@
+{"name": "Проверка вопросов и ответов на главной", "status": "passed", "description": "Проверка, что ответы соответствуют заявленным", "steps": [{"name": "Нажать на кнопку «Да все привыкли»", "status": "passed", "steps": [{"name": "Нажать на элемент", "status": "passed", "parameters": [{"name": "locator", "value": "['id', 'rcc-confirm-button']"}], "start": 1716199852050, "stop": 1716199852310}], "start": 1716199852050, "stop": 1716199852310}, {"name": "Проскроллить до блока «Вопросы о важном»", "status": "passed", "steps": [{"name": "Проскроллить до элемента", "status": "passed", "parameters": [{"name": "locator", "value": "['xpath', \".//div[text()='Вопросы о важном']\"]"}], "start": 1716199852310, "stop": 1716199852318}, {"name": "Подождать отображение элемента", "status": "passed", "parameters": [{"name": "locator", "value": "['xpath', \".//div[text()='Вопросы о важном']\"]"}], "start": 1716199852318, "stop": 1716199852330}], "start": 1716199852310, "stop": 1716199852330}, {"name": "Нажать на вопрос", "status": "passed", "steps": [{"name": "Нажать на элемент", "status": "passed", "parameters": [{"name": "locator", "value": "('id', 'accordion__heading-7')"}], "start": 1716199852330, "stop": 1716199852560}, {"name": "Подождать отображение элемента", "status": "passed", "parameters": [{"name": "locator", "value": "('id', 'accordion__heading-7')"}], "start": 1716199852560, "stop": 1716199852575}], "parameters": [{"name": "num", "value": "7"}], "start": 1716199852330, "stop": 1716199852575}, {"name": "Сравнить тексты ответов", "status": "passed", "steps": [{"name": "Подождать отображение элемента", "status": "passed", "parameters": [{"name": "locator", "value": "('id', 'accordion__panel-7')"}], "start": 1716199852575, "stop": 1716199852585}], "parameters": [{"name": "num", "value": "7"}], "start": 1716199852575, "stop": 1716199852594}], "parameters": [{"name": "num", "value": "7"}, {"name": "result", "value": "'Да, обязательно. Всем самокатов! И Москве, и Московской области.'"}], "start": 1716199852050, "stop": 1716199852595, "uuid": "3d570bd2-b010-4f32-b311-dceffc0cabbe", "historyId": "93925af55f3a914aaccc0630fddd53e7", "testCaseId": "288451571851ac6be9a051ff55691f53", "fullName": "tests.test_main_page.TestMainPage#test_check_questions_and_answers", "labels": [{"name": "parentSuite", "value": "tests"}, {"name": "suite", "value": "test_main_page"}, {"name": "subSuite", "value": "TestMainPage"}, {"name": "host", "value": "MacBook-Air-Artem.local"}, {"name": "thread", "value": "13739-MainThread"}, {"name": "framework", "value": "pytest"}, {"name": "language", "value": "cpython3"}, {"name": "package", "value": "tests.test_main_page"}]}
