+++ allure_results/31db71aa-5b45-4f6a-a7ff-a93fea79fa61-result.json
@@ -1,0 +1,1 @@
+{"name": "Проверка вопросов и ответов на главной", "status": "passed", "description": "Проверка, что ответы соответствуют заявленным", "steps": [{"name": "Нажать на кнопку «Да все привыкли»", "status": "passed", "steps": [{"name": "Нажать на элемент", "status": "passed", "parameters": [{"name": "locator", "value": "['id', 'rcc-confirm-button']"}], "start": 1716199842176, "stop": 1716199842411}], "start": 1716199842176, "stop": 1716199842411}, {"name": "Проскроллить до блока «Вопросы о важном»", "status": "passed", "steps": [{"name": "Проскроллить до элемента", "status": "passed", "parameters": [{"name": "locator", "value": "['xpath', \".//div[text()='Вопросы о важном']\"]"}], "start": 1716199842412, "stop": 1716199842418}, {"name": "Подождать отображение элемента", "status": "passed", "parameters": [{"name": "locator", "value": "['xpath', \".//div[text()='Вопросы о важном']\"]"}], "start": 1716199842418, "stop": 1716199842429}], "start": 1716199842412, "stop": 1716199842429}, {"name": "Нажать на вопрос", "status": "passed", "steps": [{"name": "Нажать на элемент", "status": "passed", "parameters": [{"name": "locator", "value": "('id', 'accordion__heading-4')"}], "start": 1716199842429, "stop": 1716199842658}, {"name": "Подождать отображение элемента", "status": "passed", "parameters": [{"name": "locator", "value": "('id', 'accordion__heading-4')"}], "start": 1716199842658, "stop": 1716199842673}], "parameters": [{"name": "num", "value": "4"}], "start": 1716199842429, "stop": 1716199842673}, {"name": "Сравнить тексты ответов", "status": "passed", "steps": [{"name": "Подождать отображение элемента", "status": "passed", "parameters": [{"name": "locator", "value": "('id', 'accordion__panel-4')"}], "start": 1716199842673, "stop": 1716199842683}], "parameters": [{"name": "num", "value": "4"}], "start": 1716199842673, "stop": 1716199842692}], "parameters": [{"name": "num", "value": "4"}, {"name": "result", "value": "'Пока что нет! Но если что-то срочное — всегда можно позвонить в поддержку по красивому номеру 1010.'"}], "start": 1716199842176, "stop": 1716199842692, "uuid": "88f2273c-402f-4d22-a318-46e06cb5b44b", "historyId": "913dc3251095e7f9473e3bda145a33c8", "testCaseId": "288451571851ac6be9a051ff55691f53", "fullName": "tests.test_main_page.TestMainPage#test_check_questions_and_answers", "labels": [{"name": "parentSuite", "value": "tests"}, {"name": "suite", "value": "test_main_page"}, {"name": "subSuite", "value": "TestMainPage"}, {"name": "host", "value": "MacBook-Air-Artem.local"}, {"name": "thread", "value": "13739-MainThread"}, {"name": "framework", "value": "pytest"}, {"name": "language", "value": "cpython3"}, {"name": "package", "value": "tests.test_main_page"}]}
