+++ allure_results/376a6026-7dbe-4a1f-8fb2-a41baceaa8ef-container.json
@@ -1,0 +1,1 @@
+{"uuid": "d5d62656-03a9-479b-b8e4-55e0a0c0648b", "befores": [{"name": "num", "status": "passed", "start": 1716199845338, "stop": 1716199845338}], "start": 1716199845338, "stop": 1716199845845}
