+++ allure_results/5f8a2131-9f50-49ed-b431-9274626e6d0e-result.json
@@ -1,0 +1,1 @@
+{"name": "Проверка вопросов и ответов на главной", "status": "passed", "description": "Проверка, что ответы соответствуют заявленным", "steps": [{"name": "Нажать на кнопку «Да все привыкли»", "status": "passed", "steps": [{"name": "Нажать на элемент", "status": "passed", "parameters": [{"name": "locator", "value": "['id', 'rcc-confirm-button']"}], "start": 1716199845339, "stop": 1716199845566}], "start": 1716199845339, "stop": 1716199845566}, {"name": "Проскроллить до блока «Вопросы о важном»", "status": "passed", "steps": [{"name": "Проскроллить до элемента", "status": "passed", "parameters": [{"name": "locator", "value": "['xpath', \".//div[text()='Вопросы о важном']\"]"}], "start": 1716199845567, "stop": 1716199845573}, {"name": "Подождать отображение элемента", "status": "passed", "parameters": [{"name": "locator", "value": "['xpath', \".//div[text()='Вопросы о важном']\"]"}], "start": 1716199845573, "stop": 1716199845589}], "start": 1716199845566, "stop": 1716199845589}, {"name": "Нажать на вопрос", "status": "passed", "steps": [{"name": "Нажать на элемент", "status": "passed", "parameters": [{"name": "locator", "value": "('id', 'accordion__heading-5')"}], "start": 1716199845589, "stop": 1716199845812}, {"name": "Подождать отображение элемента", "status": "passed", "parameters": [{"name": "locator", "value": "('id', 'accordion__heading-5')"}], "start": 1716199845812, "stop": 1716199845825}], "parameters": [{"name": "num", "value": "5"}], "start": 1716199845589, "stop": 1716199845825}, {"name": "Сравнить тексты ответов", "status": "passed", "steps": [{"name": "Подождать отображение элемента", "status": "passed", "parameters": [{"name": "locator", "value": "('id', 'accordion__panel-5')"}], "start": 1716199845825, "stop": 1716199845835}], "parameters": [{"name": "num", "value": "5"}], "start": 1716199845825, "stop": 1716199845844}], "parameters": [{"name": "num", "value": "5"}, {"name": "result", "value": "'Самокат приезжает к вам с полной зарядкой. Этого хватает на восемь суток — даже если будете кататься без передышек и во сне. Зарядка не понадобится.'"}], "start": 1716199845339, "stop": 1716199845844, "uuid": "8b601211-1497-4a50-a436-95053b432d7f", "historyId": "f01122ce1a1bb7a19bfa67500c1cf5f4", "testCaseId": "288451571851ac6be9a051ff55691f53", "fullName": "tests.test_main_page.TestMainPage#test_check_questions_and_answers", "labels": [{"name": "parentSuite", "value": "tests"}, {"name": "suite", "value": "test_main_page"}, {"name": "subSuite", "value": "TestMainPage"}, {"name": "host", "value": "MacBook-Air-Artem.local"}, {"name": "thread", "value": "13739-MainThread"}, {"name": "framework", "value": "pytest"}, {"name": "language", "value": "cpython3"}, {"name": "package", "value": "tests.test_main_page"}]}
