+++ allure_results/60d9dcd3-e1d0-4efb-ad98-c9e0879a4a93-container.json
@@ -1,0 +1,1 @@
+{"uuid": "1ca5f750-dabe-40f6-a920-7ac9e40e1226", "befores": [{"name": "num", "status": "passed", "start": 1716199828262, "stop": 1716199828262}], "start": 1716199828262, "stop": 1716199828789}
