+++ allure_results/67e1f505-ad65-4518-a51d-80a03839b62b-container.json
@@ -1,0 +1,1 @@
+{"uuid": "fef4c27c-0809-4ccb-8be8-9ecaca5a0798", "befores": [{"name": "result", "status": "passed", "start": 1716199845338, "stop": 1716199845338}], "start": 1716199845338, "stop": 1716199845844}
