+++ allure_results/6856a8af-2e60-426d-b03e-caeb3aed72d2-container.json
@@ -1,0 +1,1 @@
+{"uuid": "126faf14-ebd8-40bb-a84a-48f66679f1ef", "children": ["68d46db0-263a-4888-a23e-ae94b2c57556"], "befores": [{"name": "browser", "status": "passed", "start": 1716199836106, "stop": 1716199838869}], "afters": [{"name": "browser::0", "status": "passed", "start": 1716199839411, "stop": 1716199839789}], "start": 1716199836106, "stop": 1716199839789}
