+++ allure_results/6c6ce819-5a16-4090-bb9c-744ea67346c0-container.json
@@ -1,0 +1,1 @@
+{"uuid": "77a740df-617b-4c19-a7bb-b9942872b102", "children": ["3d570bd2-b010-4f32-b311-dceffc0cabbe"], "befores": [{"name": "browser", "status": "passed", "start": 1716199849628, "stop": 1716199852049}], "afters": [{"name": "browser::0", "status": "passed", "start": 1716199852596, "stop": 1716199852976}], "start": 1716199849628, "stop": 1716199852976}
