+++ allure_results/6d34fa63-3597-442f-a76e-3a8f9d3c780b-container.json
@@ -1,0 +1,1 @@
+{"uuid": "932a0f28-2519-4363-9e4b-c1b0c759372f", "befores": [{"name": "num", "status": "passed", "start": 1716199838869, "stop": 1716199838869}], "start": 1716199838869, "stop": 1716199839411}
