+++ allure_results/79f48376-cafb-49a8-93ea-a9011f61c4e0-container.json
@@ -1,0 +1,1 @@
+{"uuid": "396709f4-bf9a-4b7a-a635-fd32f90a8051", "children": ["2750e31b-53b0-4e97-8f94-a848cc52d96d"], "befores": [{"name": "browser", "status": "passed", "start": 1716199818490, "stop": 1716199820933}], "afters": [{"name": "browser::0", "status": "passed", "start": 1716199822268, "stop": 1716199822764}], "start": 1716199818490, "stop": 1716199822764}
