+++ allure_results/8f14985f-76c8-418f-8533-767c624d930c-container.json
@@ -1,0 +1,1 @@
+{"uuid": "a2a241a3-b297-44c9-b8bd-8eb11747734e", "befores": [{"name": "result", "status": "passed", "start": 1716199852049, "stop": 1716199852049}], "start": 1716199852049, "stop": 1716199852595}
