+++ allure_results/9311792f-72c0-4ef0-96f5-5807179c692c-container.json
@@ -1,0 +1,1 @@
+{"uuid": "714a6e4c-849e-4c34-93af-c7486d2e0cbf", "children": ["277538a9-732f-4eb3-a640-b8034bbdc321"], "befores": [{"name": "browser", "status": "passed", "start": 1716199852978, "stop": 1716199855358}], "afters": [{"name": "browser::0", "status": "passed", "start": 1716199859327, "stop": 1716199859718}], "start": 1716199852978, "stop": 1716199859719}
