+++ allure_results/97d26633-5499-4eab-94e6-b6ad4f99cebd-result.json
@@ -1,0 +1,1 @@
+{"name": "Проверка вопросов и ответов на главной", "status": "passed", "description": "Проверка, что ответы соответствуют заявленным", "steps": [{"name": "Нажать на кнопку «Да все привыкли»", "status": "passed", "steps": [{"name": "Нажать на элемент", "status": "passed", "parameters": [{"name": "locator", "value": "['id', 'rcc-confirm-button']"}], "start": 1716199828262, "stop": 1716199828498}], "start": 1716199828262, "stop": 1716199828498}, {"name": "Проскроллить до блока «Вопросы о важном»", "status": "passed", "steps": [{"name": "Проскроллить до элемента", "status": "passed", "parameters": [{"name": "locator", "value": "['xpath', \".//div[text()='Вопросы о важном']\"]"}], "start": 1716199828498, "stop": 1716199828504}, {"name": "Подождать отображение элемента", "status": "passed", "parameters": [{"name": "locator", "value": "['xpath', \".//div[text()='Вопросы о важном']\"]"}], "start": 1716199828504, "stop": 1716199828515}], "start": 1716199828498, "stop": 1716199828515}, {"name": "Нажать на вопрос", "status": "passed", "steps": [{"name": "Нажать на элемент", "status": "passed", "parameters": [{"name": "locator", "value": "('id', 'accordion__heading-0')"}], "start": 1716199828515, "stop": 1716199828755}, {"name": "Подождать отображение элемента", "status": "passed", "parameters": [{"name": "locator", "value": "('id', 'accordion__heading-0')"}], "start": 1716199828755, "stop": 1716199828770}], "parameters": [{"name": "num", "value": "0"}], "start": 1716199828515, "stop": 1716199828770}, {"name": "Сравнить тексты ответов", "status": "passed", "steps": [{"name": "Подождать отображение элемента", "status": "passed", "parameters": [{"name": "locator", "value": "('id', 'accordion__panel-0')"}], "start": 1716199828770, "stop": 1716199828780}], "parameters": [{"name": "num", "value": "0"}], "start": 1716199828770, "stop": 1716199828788}], "parameters": [{"name": "num", "value": "0"}, {"name": "result", "value": "'Сутки — 400 рублей. Оплата курьеру — наличными или картой.'"}], "start": 1716199828262, "stop": 1716199828788, "uuid": "66970d53-10a6-4970-9c15-6ed2ca6feb1f", "historyId": "83b06080596820917fb4b7045e4cb53e", "testCaseId": "288451571851ac6be9a051ff55691f53", "fullName": "tests.test_main_page.TestMainPage#test_check_questions_and_answers", "labels": [{"name": "parentSuite", "value": "tests"}, {"name": "suite", "value": "test_main_page"}, {"name": "subSuite", "value": "TestMainPage"}, {"name": "host", "value": "MacBook-Air-Artem.local"}, {"name": "thread", "value": "13739-MainThread"}, {"name": "framework", "value": "pytest"}, {"name": "language", "value": "cpython3"}, {"name": "package", "value": "tests.test_main_page"}]}
