+++ allure_results/9d9142d7-1575-403c-b555-72a691ad52dc-container.json
@@ -1,0 +1,1 @@
+{"uuid": "a714a585-ccf1-45e2-96b6-e1eb6e377df3", "befores": [{"name": "result", "status": "passed", "start": 1716199848725, "stop": 1716199848725}], "start": 1716199848725, "stop": 1716199849254}
