+++ allure_results/a0de707a-60ad-4966-99fe-5f50dd3bcd54-container.json
@@ -1,0 +1,1 @@
+{"uuid": "7cb72e22-2740-44c7-a5d9-409cc8d37e6a", "children": ["5816499a-bd59-4ca0-aa49-3bd73e144ff5"], "befores": [{"name": "browser", "status": "passed", "start": 1716199829170, "stop": 1716199831662}], "afters": [{"name": "browser::0", "status": "passed", "start": 1716199832200, "stop": 1716199832577}], "start": 1716199829170, "stop": 1716199832577}
