+++ allure_results/a1c7cfc9-94eb-48bb-9977-502e7c0a77d3-container.json
@@ -1,0 +1,1 @@
+{"uuid": "76fc59d2-eb92-4a3c-8a5c-2326fbf9e003", "children": ["88f2273c-402f-4d22-a318-46e06cb5b44b"], "befores": [{"name": "browser", "status": "passed", "start": 1716199839791, "stop": 1716199842175}], "afters": [{"name": "browser::0", "status": "passed", "start": 1716199842694, "stop": 1716199843070}], "start": 1716199839791, "stop": 1716199843070}
