+++ allure_results/a39947ce-0ae4-4fe2-9664-6060523c3a5a-container.json
@@ -1,0 +1,1 @@
+{"uuid": "543ec73d-f7eb-4bfd-844c-3e5d6a3fe132", "befores": [{"name": "num", "status": "passed", "start": 1716199852049, "stop": 1716199852049}], "start": 1716199852049, "stop": 1716199852596}
