+++ allure_results/a8fbc9cf-0b5b-4893-ab44-359bd5c35438-result.json
@@ -1,0 +1,1 @@
+{"name": "Клик по лого «Самокат»", "status": "passed", "description": "Проверка, что клик по лого «Самокат» перенаправляет на главную страницу", "steps": [{"name": "Нажать на кнопку «Заказать»", "status": "passed", "steps": [{"name": "Нажать на элемент", "status": "passed", "parameters": [{"name": "locator", "value": "['xpath', \".//div[contains(@class, 'Header_Nav')]/button[text()='Заказать']\"]"}], "start": 1716199825066, "stop": 1716199825308}], "start": 1716199825066, "stop": 1716199825308}, {"name": "Нажать на лого «Самокат»", "status": "passed", "steps": [{"name": "Нажать на элемент", "status": "passed", "parameters": [{"name": "locator", "value": "['xpath', \".//img[@alt='Scooter']\"]"}], "start": 1716199825308, "stop": 1716199825548}], "start": 1716199825308, "stop": 1716199825548}, {"name": "Проверить, что перешли на главную страницу", "status": "passed", "steps": [{"name": "Подождать отображение элемента", "status": "passed", "parameters": [{"name": "locator", "value": "['xpath', \".//div[contains(@class, 'Home_Header')]\"]"}], "start": 1716199825549, "stop": 1716199825562}], "start": 1716199825548, "stop": 1716199825562}], "start": 1716199825065, "stop": 1716199825562, "uuid": "aebf1674-dbe1-414d-b30a-32948afe6c2f", "historyId": "ebcdefb6a19fa49fe3552707dac79acd", "testCaseId": "ebcdefb6a19fa49fe3552707dac79acd", "fullName": "tests.test_main_page.TestMainPage#test_click_logo_scooter_follow_link_main", "labels": [{"name": "parentSuite", "value": "tests"}, {"name": "suite", "value": "test_main_page"}, {"name": "subSuite", "value": "TestMainPage"}, {"name": "host", "value": "MacBook-Air-Artem.local"}, {"name": "thread", "value": "13739-MainThread"}, {"name": "framework", "value": "pytest"}, {"name": "language", "value": "cpython3"}, {"name": "package", "value": "tests.test_main_page"}]}
