+++ allure_results/b141ebab-871a-4b47-a7bc-fccb1f8b35b6-container.json
@@ -1,0 +1,1 @@
+{"uuid": "932438d0-0789-44d7-930c-3a593cf66e23", "children": ["f31f5fcc-0680-4912-a38f-ec658a89a725"], "befores": [{"name": "browser", "status": "passed", "start": 1716199832579, "stop": 1716199835194}], "afters": [{"name": "browser::0", "status": "passed", "start": 1716199835728, "stop": 1716199836104}], "start": 1716199832579, "stop": 1716199836104}
