+++ allure_results/b1490029-b47d-4854-8117-aa25cbe28c8c-container.json
@@ -1,0 +1,1 @@
+{"uuid": "e53e6fb8-515d-4639-aeb0-01d7766a3ed1", "children": ["aebf1674-dbe1-414d-b30a-32948afe6c2f"], "befores": [{"name": "browser", "status": "passed", "start": 1716199822765, "stop": 1716199825065}], "afters": [{"name": "browser::0", "status": "passed", "start": 1716199825563, "stop": 1716199825950}], "start": 1716199822765, "stop": 1716199825950}
