+++ allure_results/b473a958-4caa-4402-addb-808a85b2d7bc-result.json
@@ -1,0 +1,1 @@
+{"name": "Проверка вопросов и ответов на главной", "status": "passed", "description": "Проверка, что ответы соответствуют заявленным", "steps": [{"name": "Нажать на кнопку «Да все привыкли»", "status": "passed", "steps": [{"name": "Нажать на элемент", "status": "passed", "parameters": [{"name": "locator", "value": "['id', 'rcc-confirm-button']"}], "start": 1716199848726, "stop": 1716199848975}], "start": 1716199848726, "stop": 1716199848975}, {"name": "Проскроллить до блока «Вопросы о важном»", "status": "passed", "steps": [{"name": "Проскроллить до элемента", "status": "passed", "parameters": [{"name": "locator", "value": "['xpath', \".//div[text()='Вопросы о важном']\"]"}], "start": 1716199848975, "stop": 1716199848979}, {"name": "Подождать отображение элемента", "status": "passed", "parameters": [{"name": "locator", "value": "['xpath', \".//div[text()='Вопросы о важном']\"]"}], "start": 1716199848979, "stop": 1716199848988}], "start": 1716199848975, "stop": 1716199848988}, {"name": "Нажать на вопрос", "status": "passed", "steps": [{"name": "Нажать на элемент", "status": "passed", "parameters": [{"name": "locator", "value": "('id', 'accordion__heading-6')"}], "start": 1716199848988, "stop": 1716199849226}, {"name": "Подождать отображение элемента", "status": "passed", "parameters": [{"name": "locator", "value": "('id', 'accordion__heading-6')"}], "start": 1716199849226, "stop": 1716199849239}], "parameters": [{"name": "num", "value": "6"}], "start": 1716199848988, "stop": 1716199849239}, {"name": "Сравнить тексты ответов", "status": "passed", "steps": [{"name": "Подождать отображение элемента", "status": "passed", "parameters": [{"name": "locator", "value": "('id', 'accordion__panel-6')"}], "start": 1716199849239, "stop": 1716199849247}], "parameters": [{"name": "num", "value": "6"}], "start": 1716199849239, "stop": 1716199849253}], "parameters": [{"name": "num", "value": "6"}, {"name": "result", "value": "'Да, пока самокат не привезли. Штрафа не будет, объяснительной записки тоже не попросим. Все же свои.'"}], "start": 1716199848726, "stop": 1716199849253, "uuid": "1b07fbd2-d8d0-4816-9a7a-a77ff89255b4", "historyId": "7950ffdb4ebbdecdf431874469d27170", "testCaseId": "288451571851ac6be9a051ff55691f53", "fullName": "tests.test_main_page.TestMainPage#test_check_questions_and_answers", "labels": [{"name": "parentSuite", "value": "tests"}, {"name": "suite", "value": "test_main_page"}, {"name": "subSuite", "value": "TestMainPage"}, {"name": "host", "value": "MacBook-Air-Artem.local"}, {"name": "thread", "value": "13739-MainThread"}, {"name": "framework", "value": "pytest"}, {"name": "language", "value": "cpython3"}, {"name": "package", "value": "tests.test_main_page"}]}
