+++ allure_results/d238e810-b007-41c4-9042-a8e5c1a13659-container.json
@@ -1,0 +1,1 @@
+{"uuid": "43f9bec4-4efc-4dfb-ba8c-9e1a38d89903", "befores": [{"name": "result", "status": "passed", "start": 1716199831663, "stop": 1716199831663}], "start": 1716199831663, "stop": 1716199832199}
