+++ allure_results/dd9375bc-8e9a-43eb-a057-7d4ac8ceff75-result.json
@@ -1,0 +1,1 @@
+{"name": "Проверка вопросов и ответов на главной", "status": "passed", "description": "Проверка, что ответы соответствуют заявленным", "steps": [{"name": "Нажать на кнопку «Да все привыкли»", "status": "passed", "steps": [{"name": "Нажать на элемент", "status": "passed", "parameters": [{"name": "locator", "value": "['id', 'rcc-confirm-button']"}], "start": 1716199838871, "stop": 1716199839123}], "start": 1716199838871, "stop": 1716199839123}, {"name": "Проскроллить до блока «Вопросы о важном»", "status": "passed", "steps": [{"name": "Проскроллить до элемента", "status": "passed", "parameters": [{"name": "locator", "value": "['xpath', \".//div[text()='Вопросы о важном']\"]"}], "start": 1716199839124, "stop": 1716199839131}, {"name": "Подождать отображение элемента", "status": "passed", "parameters": [{"name": "locator", "value": "['xpath', \".//div[text()='Вопросы о важном']\"]"}], "start": 1716199839132, "stop": 1716199839151}], "start": 1716199839124, "stop": 1716199839151}, {"name": "Нажать на вопрос", "status": "passed", "steps": [{"name": "Нажать на элемент", "status": "passed", "parameters": [{"name": "locator", "value": "('id', 'accordion__heading-3')"}], "start": 1716199839152, "stop": 1716199839376}, {"name": "Подождать отображение элемента", "status": "passed", "parameters": [{"name": "locator", "value": "('id', 'accordion__heading-3')"}], "start": 1716199839376, "stop": 1716199839391}], "parameters": [{"name": "num", "value": "3"}], "start": 1716199839152, "stop": 1716199839391}, {"name": "Сравнить тексты ответов", "status": "passed", "steps": [{"name": "Подождать отображение элемента", "status": "passed", "parameters": [{"name": "locator", "value": "('id', 'accordion__panel-3')"}], "start": 1716199839391, "stop": 1716199839401}], "parameters": [{"name": "num", "value": "3"}], "start": 1716199839391, "stop": 1716199839410}], "parameters": [{"name": "num", "value": "3"}, {"name": "result", "value": "'Только начиная с завтрашнего дня. Но скоро станем расторопнее.'"}], "start": 1716199838870, "stop": 1716199839410, "uuid": "68d46db0-263a-4888-a23e-ae94b2c57556", "historyId": "c41a574fd3bc6191b6a44e54ffc99180", "testCaseId": "288451571851ac6be9a051ff55691f53", "fullName": "tests.test_main_page.TestMainPage#test_check_questions_and_answers", "labels": [{"name": "parentSuite", "value": "tests"}, {"name": "suite", "value": "test_main_page"}, {"name": "subSuite", "value": "TestMainPage"}, {"name": "host", "value": "MacBook-Air-Artem.local"}, {"name": "thread", "value": "13739-MainThread"}, {"name": "framework", "value": "pytest"}, {"name": "language", "value": "cpython3"}, {"name": "package", "value": "tests.test_main_page"}]}
