+++ allure_results/e70dcf36-1056-4abb-8dc4-1aed0dcc5e99-container.json
@@ -1,0 +1,1 @@
+{"uuid": "6bb6e64b-0798-4ef4-acd2-2d778e4422af", "befores": [{"name": "num", "status": "passed", "start": 1716199835194, "stop": 1716199835194}], "start": 1716199835194, "stop": 1716199835728}
