+++ allure_results/ea070b03-abf1-4e33-8bad-ac094e65279f-container.json
@@ -1,0 +1,1 @@
+{"uuid": "2959bbf7-b4ab-4de2-b540-26c0c16c0461", "befores": [{"name": "result", "status": "passed", "start": 1716199828262, "stop": 1716199828262}], "start": 1716199828262, "stop": 1716199828789}
