+++ allure_results/fe882727-1ae7-4218-bccb-317cc4a6992b-container.json
@@ -1,0 +1,1 @@
+{"uuid": "2b8bce11-7e03-4ce7-b59b-78aa1999e32a", "befores": [{"name": "result", "status": "passed", "start": 1716199842176, "stop": 1716199842176}], "start": 1716199842176, "stop": 1716199842692}
