+++ allure_results/ffd2753a-4ebf-4a8b-bf85-eeccf3213976-container.json
@@ -1,0 +1,1 @@
+{"uuid": "51d9670e-4c57-43ed-9a12-48850fc97994", "children": ["8b601211-1497-4a50-a436-95053b432d7f"], "befores": [{"name": "browser", "status": "passed", "start": 1716199843072, "stop": 1716199845338}], "afters": [{"name": "browser::0", "status": "passed", "start": 1716199845845, "stop": 1716199846224}], "start": 1716199843072, "stop": 1716199846224}
