+++ data.py
@@ -1,12 +1,6 @@
 import random
 
 
-class Urls:
-    MAIN_URL = "https://qa-scooter.praktikum-services.ru/"  # Адрес главной страницы
-    DZEN_URL = "https://dzen.ru/?yredirect=true"  # Адрес страницы «Дзен»
-    ORDER_URL = "https://qa-scooter.praktikum-services.ru/order"  # Адрес страницы заказа
-
-
 class Data:
     NAMES = ["Артем", "Ангелина", "Света", "Владимир", "Роман", "Андрей"]
     NAME = random.choice(NAMES)
@@ -14,7 +8,8 @@ class Data:
     LASTNAMES = ["Стопов", "Разумовская", "Бушев", "Медведева", "Крылатов", "Соболев"]
     LASTNAME = random.choice(LASTNAMES)
 
-    ADDRESSES = ["Санкт-Петербург, Невский проспект, 85", "Москва, Перовская улица 10", "Муром, Комсомольская улица, 51"]
+    ADDRESSES = ["Санкт-Петербург, Невский проспект, 85", "Москва, Перовская улица 10",
+                 "Муром, Комсомольская улица, 51"]
     ADDRESS = random.choice(ADDRESSES)
 
     PHONE = f'+7{random.randint(9000000000, 9999999999)}'
