+++ tests/conftest.py
@@ -1,15 +1,11 @@
 import pytest
 from selenium import webdriver
 
-from data import Urls
-
-urls = Urls()
-
 
 @pytest.fixture
 def browser():
     driver = webdriver.Firefox()
-    driver.get(urls.MAIN_URL)
+    driver.get("https://qa-scooter.praktikum-services.ru/")
 
     yield driver
     driver.quit()
